bento-cluster: CLUSTER-10. bento-cluster should print ports of services started.
This commit makes two modifications to bento-cluster:

1. bento-cluster now reports on the ports the NameNode, JobTracker, and Zookeeper
services are available on.
2. Ports / web addresses are reported both when the cluster is configured and when
the cluster starts.

To facilitate this, a new tool, PortReportingTool, has been created that can report
on the ports being used by bento-cluster. This tool runs after the cluster is configured
or the cluster is started. The new tool was needed to report on ports after the cluster
starts, because if this code was in MiniClusterTool itself, it would be gobbled up by
the bento script's redirection of cluster output.

Signed-off-by: Aaron Kimball <aaron@wibidata.com>
commit: https://github.com/kijiproject/bento-cluster/commit/d7672f73601f061aca1172160716c6bb7cbc9e61