kiji-express: EXP-189. Add a cleanup method to the score producer.
There are tables and resources that are opened during the setup of the producer
but not closed hence throwing exceptions in the finalizer if the producer is in
a long running process. This adds a cleanup method to close and release
resources.

Reviewer: robert
commit: https://github.com/kijiproject/kiji-express/commit/02b8d8e89707a696debf352325278ee19ccd92cb